requirements.txt
azure-devops==6.0.0b2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isodate==0.6.0
msrest==0.6.11
oauthlib==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
